Module Ocaml_protoc_compiler_lib.Pb_codegen_backend

Compile protobuf parse tree to the ocaml type

This module focuses on the compilation steps which transforms a fully resolved Protobuf message into an OCaml representation.

After compilation this module also expose code generation functionality.

module Tt = Pb_typing_type_tree

module Ot = Pb_codegen_ocaml_type

Compilation

val compile : unsigned_tag:bool -> Pb_field_type.resolved Tt.proto -> Pb_field_type.resolved Tt.proto_type -> Ot.type_ list

module Internal : sig ... end

Internal helpers.