Module Pb_codegen_backend.Internal

Internal helpers.

For now there are no guarantees of stability within this module. It's unlikely that the signatures will change, but the possibility is still there.

val is_mutable : ?field_name:string -> Pb_option.set -> bool

val constructor_name : string -> string

val module_name : string -> string

val label_name_of_field_name : string -> string

val type_name : string list -> string -> string

val variant_of_oneof : ?include_oneof_name:unit -> outer_message_names:string list -> unsigned_tag:bool -> 'a Tt.proto -> Pb_option.set -> string -> int Tt.oneof -> Ot.variant