Module Ocaml_protoc_compiler_lib.Pb_codegen_decode_yojson

Code generator for the decode JSON function

include Pb_codegen_sig.S

val gen_sig : ?and_:unit -> Pb_codegen_ocaml_type.type_ -> Pb_codegen_formatting.scope -> bool

gen_sig t generates the code for the module signature

val gen_struct : ?and_:unit -> Pb_codegen_ocaml_type.type_ -> Pb_codegen_formatting.scope -> bool

gen_struct t generates the code for the module structure

val ocamldoc_title : string

OCamldoc title

val file_suffix : string

The suffix part of the generated file which will contain the struct and sig