Module Ocaml_protoc_compiler_lib.Pb_codegen_plugin

Plugin architecture.

OCaml-protoc generates code based on a number of plugins, each of which can contribute code to the output files (.ml and .mli).

type codegen_f = ?and_:unit -> Pb_codegen_ocaml_type.type_ -> Pb_codegen_formatting.scope -> bool

A code generation function

module type S = sig ... end

type t = (module S)

A plugin is a code-generator respecting the signature S.

val requires_mutable_records : (module S) -> bool