Module Ocaml_protoc_compiler_lib.Pb_parsing_parse_tree

Protobuf parse tree

type message_field_label = [

| `Optional

| `Required

| `Repeated

| `Nolabel

]

A field property defining its occurence

type oneof_field_label = unit

Oneof field fields label

Oneof fields have no label, they are simply choices for the oneof fiel they belong to.

type 'a field = {

field_name : string;

field_number : int;

field_label : 'a;

field_type : Pb_field_type.unresolved_t;

field_options : Pb_option.set;

}

message field.

Note this field is parametrized with the label type so that it can be used both by normal field and one of field since the only difference between the 2 is the label.

type message_field = message_field_label field

type oneof_field = oneof_field_label field

type map_field = {

map_name : string;

map_number : int;

map_key_type : Pb_field_type.map_key_type;

map_value_type : Pb_field_type.unresolved_t;

map_options : Pb_option.set;

}

type oneof_body_content =

| Oneof_field of oneof_field

| Oneof_option of Pb_option.t

type oneof = {

oneof_name : string;

oneof_body : oneof_body_content list;

}

oneof entity

type enum_value = {

enum_value_name : string;

enum_value_int : int;

}

type enum_body_content =

| Enum_value of enum_value

| Enum_option of Pb_option.t

type enum = {

enum_id : int;

enum_name : string;

enum_body : enum_body_content list;

}

type extension_range_to =

| To_max

| To_number of int

type extension_range_from = int

type extension_range =

| Extension_single_number of int

| Extension_range of extension_range_from * extension_range_to

type message_body_content =

| Message_field of message_field

| Message_map_field of map_field

| Message_oneof_field of oneof

| Message_sub of message

| Message_enum of enum

| Message_extension of extension_range list

| Message_reserved of extension_range list

| Message_option of Pb_option.t

Body content defines all the possible consituant of a message.

and message = {

id : int;

message_name : string;

message_body : message_body_content list;

}

Message entity.

Note the ID is simply for uniquely (and easily) identifying a type. It is expected to be generated by a parser. The later compilation functions expects this id to be unique.

type rpc = {

rpc_name : string;

rpc_options : Pb_option.set;

rpc_req_stream : bool;

rpc_req : Pb_field_type.unresolved_t;

rpc_res_stream : bool;

rpc_res : Pb_field_type.unresolved_t;

}

type service_body_content =

| Service_rpc of rpc

| Service_option of Pb_option.t

type service = {

service_name : string;

service_body : service_body_content list;

}

type extend = {

id : int;

extend_name : string;

extend_body : message_field list;

}

type import = {

file_name : string;

public : bool;

}

type proto = {

proto_file_name : string option;

syntax : string option;

imports : import list;

file_options : Pb_option.set;

package : string option;

messages : message list;

services : service list;

enums : enum list;

extends : extend list;

}

Definition of a protobuffer message file.

val pp_message_field_label : Stdlib.Format.formatter -> [< `Nolabel | `Optional | `Repeated | `Required ] -> unit

val pp_oneof_field_label : 'a -> unit -> unit

val pp_field : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'b field -> unit

val pp_message_field : Stdlib.Format.formatter -> [< `Nolabel | `Optional | `Repeated | `Required ] field -> unit

val pp_oneof_field : Stdlib.Format.formatter -> unit field -> unit

val pp_map_field : Stdlib.Format.formatter -> map_field -> unit

val pp_oneof_body_content : Stdlib.Format.formatter -> oneof_body_content -> unit

val pp_oneof : Stdlib.Format.formatter -> oneof -> unit

val pp_enum_value : Stdlib.Format.formatter -> enum_value -> unit

val pp_enum_body_content : Stdlib.Format.formatter -> enum_body_content -> unit

val pp_enum : Stdlib.Format.formatter -> enum -> unit

val pp_extension_range_to : Stdlib.Format.formatter -> extension_range_to -> unit

val pp_extension_range_from : Stdlib.Format.formatter -> int -> unit

val pp_extension_range : Stdlib.Format.formatter -> extension_range -> unit

val pp_message_body_content : Stdlib.Format.formatter -> message_body_content -> unit

val pp_message : Stdlib.Format.formatter -> message -> unit

val pp_rpc : Stdlib.Format.formatter -> rpc -> unit

val pp_service_body_content : Stdlib.Format.formatter -> service_body_content -> unit

val pp_service : Stdlib.Format.formatter -> service -> unit

val pp_extend : Stdlib.Format.formatter -> extend -> unit

val pp_import : Stdlib.Format.formatter -> import -> unit

val pp_proto : Stdlib.Format.formatter -> proto -> unit