Module Pb_typing_resolution.Types_by_scope

Custom container for all the types (message or enums) which are organized by their scope. This allow efficient search of a type given its type path

The construction of this container is the first step in the type resolution.

type t

container type

val empty : t

empty container

val add : t -> Pb_field_type.unresolved Tt.proto_type -> t

add a protobuf type

val find : t -> Pb_field_type.type_path -> string -> Pb_field_type.unresolved Tt.proto_type

find a protobuf type given its type path

val print : Stdlib.out_channel -> t -> unit

pretty print of the container