Module Ocaml_protoc_compiler_lib.Pb_typing_validation

Typed tree construction and validation

This module focuses on building the Type tree from the parse tree and provides a number of validation in doing so.

Not that the field types are not resolved, and the returned type tree is hence of type unresolved_field_type Pb_typing_type_tree.proto_type. The type resolution is done in Pb_typing_resolution

module Pt = Pb_parsing_parse_tree

module Tt = Pb_typing_type_tree

Type tree construction

val validate : Pt.proto -> Pb_field_type.unresolved Tt.proto

validate file_name proto makes a first phase compilation of the parsed tree.

Testing Only

val validate_message : ?parent_options:Pb_option.set -> string -> Pb_option.set -> Tt.type_scope -> Pt.message -> Pb_field_type.unresolved Tt.proto_type list