Module Ocaml_protoc_compiler_lib.Pb_util

Standard library missing functions

val rev_split_by_char : char -> string -> string list

rev_split_by_char c s will split the string s using the delimieter c and return the component in reverse order (ie from right to left).

For instance when splitting a filename with the '.' character, the file extension will be the head of the returned list.

val string_of_string_list : string list -> string

string_of_string_list l returns a debug string of l

val string_fold_lefti : ('a -> int -> char -> 'a) -> 'a -> string -> 'a

string_fold_lefti f e0 s will fold over each string character

val indentation_prefix : int -> string

indentation_prefix level returns a string of 2 * level spaces

module String : sig ... end

module Option : sig ... end

val read_file : string -> string

val protect : finally:(unit -> unit) -> (unit -> 'a) -> 'a

module List : sig ... end

module Str_map : Stdlib.Map.S with type key = string

module Int_map : Stdlib.Map.S with type key = int