Module Pb_util.List

val pop_last : 'a list -> 'a list

pop_last l removes the last element from the list

val apply_until : ('a -> 'b option) -> 'a list -> 'b option

apply_until f l applies f ei until it returns Some x

If the end of the list is reached without f returning Some _ then None is returned.

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

filter_map f l returns the list of element x for which f returned Some x. The length of the returned list will be less or equal than the length of the input list l.