Module Pb_util.Option

val default : 'a -> 'a option -> 'a

option_default x o returns x is o is None otherwise y when o is Some y.

val min_value : 'a option -> 'a option -> 'a option

min_value x y returns the min value of the x and y if x and y are of the form Some _. If x or y is None then Invalid_argument exception is raised

val eq_value : 'a option -> 'a option -> bool

eq_value x y returns true if both x and y are of the form Some _. If x or y is None then Invalid_argument exception is raised

val string_of_option : ('a -> string) -> 'a option -> string

string_of_option f x returns string representation of x using f when x is of the form Some _.