Module Ocaml_protoc_compiler_lib.Pb_codegen_ocaml_type

OCaml type representation

module Pt = Pb_parsing_parse_tree

type payload_kind =

| Pk_varint of bool
(*
zigzag
*)

| Pk_bits32

| Pk_bits64

| Pk_bytes

type user_defined_type = {

udt_module_prefix : string option;
(*
since code generated is split in multiple file (type, binary, json, ..) this defines the prefix for the given type, the suffix will be defined by each generator
*)

udt_type_name : string;
(*
OCaml type name ie not the type name in proto file
*)

udt_type : [ `Message | `Enum ];
(*
Need to keep track of this since encoding logic in binary format is quite different
*)

}

type basic_type =

| Bt_string

| Bt_float

| Bt_int

| Bt_int32

| Bt_uint32

| Bt_int64

| Bt_uint64

| Bt_bytes

| Bt_bool

type wrapper_type = {

wt_type : basic_type;

wt_pk : payload_kind;

}

type field_type =

| Ft_unit

| Ft_basic_type of basic_type

| Ft_user_defined_type of user_defined_type

| Ft_wrapper_type of wrapper_type
(*
New wrapper type which indicates that the corresponding ocaml Type should be an `option` along with the fact that it is encoded with special rules
*)

type default_value = Pb_option.constant option

type associative_type =

| At_list

| At_hashtable

type repeated_type =

| Rt_list

| Rt_repeated_field

type encoding_number = int

type is_packed = bool

type record_field_type =

| Rft_nolabel of field_type * encoding_number * payload_kind
(*
no default values in proto3 no label fields
*)

| Rft_required of field_type * encoding_number * payload_kind * default_value

| Rft_optional of field_type * encoding_number * payload_kind * default_value

| Rft_repeated of repeated_type * field_type * encoding_number * payload_kind * is_packed

| Rft_associative of associative_type * encoding_number * (basic_type * payload_kind) * (field_type * payload_kind)

| Rft_variant of variant

and variant_constructor = {

vc_constructor : string;

vc_field_type : variant_constructor_type;

vc_encoding_number : encoding_number;

vc_payload_kind : payload_kind;

vc_options : Pb_option.set;

}

and variant_constructor_type =

| Vct_nullary

| Vct_non_nullary_constructor of field_type

and variant = {

v_name : string;

v_constructors : variant_constructor list;

}

and record_field = {

rf_label : string;

rf_field_type : record_field_type;

rf_mutable : bool;

rf_options : Pb_option.set;

}

and record = {

r_name : string;

r_fields : record_field list;

}

and const_variant_constructor = {

cvc_name : string;

cvc_binary_value : int;

cvc_string_value : string;

cvc_options : Pb_option.set;

}

and const_variant = {

cv_name : string;

cv_constructors : const_variant_constructor list;

}

and empty_record = {

er_name : string;

}

and type_spec =

| Record of record

| Variant of variant

| Const_variant of const_variant

| Unit of empty_record

type type_ = {

module_prefix : string;
(*
code generation leads to several file/module being generated for a given type_. module_prefix is the common prefix for all those generated module and it is based on the `.proto` filename.
*)

spec : type_spec;

type_level_ppx_extension : string option;

type_options : Pb_option.set;

}

type rpc_type =

| Rpc_scalar of field_type

| Rpc_stream of field_type

RPC argument or return type. We require message types in RPC.

type rpc = {

rpc_name : string;

rpc_req : rpc_type;

rpc_res : rpc_type;

}

A RPC specification, ie the signature for one remote procedure.

type service = {

service_name : string;

service_packages : string list;
(*
Package in which this belongs
*)

service_body : rpc list;

}

A service, composed of multiple RPCs.

type proto = {

proto_types : type_ list list;
(*
List of strongly connected type definitions
*)

proto_services : service list;

}

A proto file is composed of a list of types and a list of services.